Only limit results in specific cases
We assume that we want to check for matches in both userid
and email. However, if the search string looks like an email
address, we check if there are multiple accounts with the same
email address registred. If so, we limit the search to userid.
We will never search the name keys.

Signed-off-by: Micke Nordin <kano@sunet.se>